refactor: moved formatter to formatting
